Merged change 7edb64f6-3e05-488f-973c-090e1af2eb42

From review branch elliott/execute_in_path into master

+++ libraries/node_install.rb
@@ -52,6 +52,7 @@ class Chef
           NODE_PATH=$( dirname $(nvm which #{new_resource.version}))
           cp -R $NODE_PATH/../../#{new_resource.version} #{prefix}/
         CODE
+        not_if { ::File.exist?("#{prefix}/#{new_resource.version}") }
       end
     end
   end
